google login is clickable, made the LoginForm a separate component, need to rework button styling
