+++ src/app/login/page.tsx
@@ -1,5 +1,6 @@
 import styles from '../../styles/login.module.css'
 import Image from 'next/image'
+import LoginForm from '@/components/LoginForm'
 
 export default function LoginPage() {
   return (
@@ -13,33 +14,7 @@ export default function LoginPage() {
         />
       </div>
       <div className={styles.formSection}>
-        <form className={styles.form}>
-          <h1 className={styles.welcomeMessage}>
-            Welcome Back to PresencePilot
-          </h1>
-
-          <button type="submit" className={styles.button}>
-            Continue with Google
-          </button>
-
-          <input
-            type="email"
-            placeholder="Email"
-            className={styles.input}
-            required
-          />
-
-          <input
-            type="password"
-            placeholder="Password"
-            className={styles.input}
-            required
-          />
-
-          <button type="submit" className={styles.button}>
-            Login
-          </button>
-        </form>
+        <LoginForm />
       </div>
     </div>
   )
